Add files via upload
So Webcam version is the correctly implemented UI as intended from sir kelvin, but i dont have the djtello drone so i cant see whether it works properly in the drone version or not. if it isnt, you can refer to the webcam version and see how they positioned the buttons, it should be good and simple.